+++ perl/FSM/HDL/FlattenedDT.pm
@@ -21,6 +21,7 @@ use FSM::Synthesis::EnableGraph::FactorizationSupport;
 use FSM::Synthesis::EnableGraph::IntermediateSignalSupport;
 use FSM::Synthesis::EnableGraph::ModulePlanningSupport;
 use FSM::Synthesis::EnableGraph::SignalSupport;
+use FSM::HDL::FlattenedDT::DecisionTreeFlatteningSupport;
 use FSM::HDL::FlattenedDT::Orchestrator;
 use FSM::HDL::FlattenedDT::Backend::SystemVerilog::ASTFactorizationSupport;
 use FSM::HDL::FlattenedDT::Backend::SystemVerilog::ConsolidatedIntermediateAssignmentSupport;
@@ -116,6 +117,7 @@ sub new ($class, %args) {
     $self->{enable_graph_intermediate_support} = FSM::Synthesis::EnableGraph::IntermediateSignalSupport->new(flattened_dt => $self);
     $self->{enable_graph_module_planning_support} = FSM::Synthesis::EnableGraph::ModulePlanningSupport->new(flattened_dt => $self);
     $self->{enable_graph_signal_support} = FSM::Synthesis::EnableGraph::SignalSupport->new(flattened_dt => $self);
+    $self->{decision_tree_flattening_support} = FSM::HDL::FlattenedDT::DecisionTreeFlatteningSupport->new(flattened_dt => $self);
     $self->{orchestrator} = FSM::HDL::FlattenedDT::Orchestrator->new(flattened_dt => $self);
     $self->{backend_sv_scaffold} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::ScaffoldEmitter->new(flattened_dt => $self);
     $self->{backend_sv_internal_decl} = FSM::HDL::FlattenedDT::Backend::SystemVerilog::InternalDeclarationEmitter->new(flattened_dt => $self);